app/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
asyncpg==0.24.0
passlib==1.7.4
psycopg2-binary==2.9.1
